HighwayMonitoring_ML/AccidentDetection_Live/requirements.txt
tensorflow==2.4.0
keras==2.4.3 
numpy==1.19.3 
pillow==7.0.0 
scipy==1.4.1 
h5py==2.10.0 
matplotlib==3.3.2 
keras-resnet==0.2.0
imageai==2.1.6
opencv_python
fastapi
uvicorn
azure-cosmos
requests==2.27.1
